entrain-web/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
rq>=1.16.0
redis>=5.0.0
elevenlabs>=1.0.0
numpy>=1.26.0
scipy>=1.12.0
soundfile>=0.12.0
pyyaml>=6.0.0
python-multipart>=0.0.6
httpx>=0.26.0
